+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 2.7.0
+
+- Adds `header.title` and `header.subtitle` YAML config to customize Cortex page title and subtitle
+- Adds `hideCortexLinks` YAML flag to hide all links to Cortex app
+- Adds `initiativeNameOverride` YAML config to override "Initiative" name
+
 ### 2.6.6
 
 - Fix rule description and rule query text colors